scripts/scraper/requirements.txt
requests>=2.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
